pipeline/cloud/services/vertex_processor/requirements.txt
vertexai
google-cloud-storage
